Deprecate importing qip functions to the qutip namespace (#1074)
* wrapper for the qip deprecation warning

* warning shows only once by repeated importation

* pep8

* Update qip_depracation.py

* move gates into ops folder for later adding measurement, classical control

* from qutip.qip.ops import instead of from qutip.qip import

* fix the trailing whitespace

* deprecation warning for qutip.qip.gates

* circular importation, use absolute module path

* seems to be a hidden circular importation

* import ops before circuit?

* call gates only when used, not at the begining

* try again

* move deprecation wrapper inside qip folder

* rename ops to operations

* add package to setup.py

* fix a typo

* fix a typo during renaming

* last try with relative importation
