+++ qutip/entropy.py
@@ -41,7 +41,6 @@ from qutip.states import ket2dm
 from qutip.tensor import tensor
 from qutip.operators import sigmay
 from qutip.sparse import sp_eigs
-from qutip.qip.gates import swap
 from qutip.partial_transpose import partial_transpose
 
 
@@ -349,6 +348,7 @@ def entangling_power(U):
     if U.dims != [[2, 2], [2, 2]]:
         raise Exception("U must be a two-qubit gate.")
 
+    from qutip.qip.operations.gates import swap
     a = (tensor(U, U).dag() * swap(N=4, targets=[1, 3]) *
          tensor(U, U) * swap(N=4, targets=[1, 3]))
     b = (tensor(swap() * U, swap() * U).dag() * swap(N=4, targets=[1, 3]) *
