+++ qutip/qip/circuit.py
@@ -38,7 +38,7 @@ import inspect
 import numpy as np
 
 from qutip.qip.circuit_latex import _latex_compile
-from qutip.qip.gates import *
+from qutip.qip.operations.gates import *
 from qutip.qip.qubits import qubit_states
 
 __all__ = ['Gate', 'QubitCircuit']
