+++ qutip/tests/test_spinchain.py
@@ -33,7 +33,8 @@
 
 import numpy as np
 from numpy.testing import assert_, run_module_suite, assert_allclose
-from qutip.qip.gates import gate_sequence_product
+
+from qutip.qip.operations.gates import gate_sequence_product
 from qutip.qip.circuit import QubitCircuit
 from qutip.qip.device.spinchain import LinearSpinChain, CircularSpinChain
 from qutip.random_objects import rand_ket
